+++ src/Client.php
@@ -10,10 +10,13 @@ use GingerPluginSdk\Entities\Order;
 use GingerPluginSdk\Exceptions\APIException;
 use GingerPluginSdk\Exceptions\CaptureFailedException;
 use GingerPluginSdk\Exceptions\InvalidOrderDataException;
+use GingerPluginSdk\Exceptions\InvalidOrderStatusException;
 use GingerPluginSdk\Exceptions\OrderNotFoundException;
+use GingerPluginSdk\Exceptions\RefundFailedException;
 use GingerPluginSdk\Helpers\HelperTrait;
 use GingerPluginSdk\Interfaces\AbstractCollectionContainerInterface;
 use GingerPluginSdk\Interfaces\ArbitraryArgumentsEntityInterface;
+use GingerPluginSdk\Interfaces\ValueInCentsInterface;
 use GingerPluginSdk\Properties\ClientOptions;
 use GingerPluginSdk\Properties\Currency;
 use RuntimeException;
@@ -59,7 +62,7 @@ class Client
      *
      * @throws \Exception
      */
-    public function getOrder(string $id): object
+    public function getOrder(string $id): Order
     {
         try {
             $api_order = $this->api_client->getOrder(
@@ -81,16 +84,16 @@ class Client
      * Only orders with supporting capturing payment methods is allowed, for example klarna-pay-later or afterpay.
      *
      * @throws \GingerPluginSdk\Exceptions\CaptureFailedException
-     * @throws \GingerPluginSdk\Exceptions\InvalidOrderDataException
+     * @throws \GingerPluginSdk\Exceptions\InvalidOrderStatusException
      */
     public function captureOrderTransaction(string $id): bool
     {
-        /** @var Order $order */
         $order = $this->getOrder(id: $id);
 
         if ($order->getStatus() !== 'completed') {
-            throw new InvalidOrderDataException(
-                message: sprintf("Only order with `completed` status could be captured, current order status is %s", $order->getStatus()));
+            throw new InvalidOrderStatusException(
+                actual: $order->getStatus(), expected: 'completed'
+            );
         }
 
         try {
@@ -118,7 +121,7 @@ class Client
      * @phpstan-param class-string<Q> $className
      * @phpstan-return Q
      */
-    public function fromArray(string $className, array $data): object
+    public function fromArray(string $className, array $data): mixed
     {
         $arguments = [];
         foreach ($data as $property_name => $value) {
@@ -150,6 +153,9 @@ class Client
                 //Check if this property has a pattern validation
                 $path_to_property = self::PROPERTIES_PATH . $this->dashesToCamelCase($property_name, true);
                 if (class_exists($path_to_property)) {
+                    if (in_array('ValueInCentsInterface', class_implements($path_to_property))) {
+                        $value /= 100;
+                    }
                     $arguments[$camel_property_name] = new $path_to_property($value);
                 } else {
                     if (array_key_exists(ArbitraryArgumentsEntityInterface::class, class_implements($className))) {
@@ -244,10 +250,39 @@ class Client
     }
 
     /**
-     * @throws \GingerPluginSdk\Exceptions\ValidationException
-     * @throws \Exception
+     * Refund order using order lines.
+     * Only completed orders could be refunded.
+     *
+     * @param string $order_id
+     * @return array
+     * @throws \GingerPluginSdk\Exceptions\InvalidOrderStatusException
+     * @throws \GingerPluginSdk\Exceptions\RefundFailedException
      */
-    public function sendOrder(Order $order): object
+    public function refundOrder(string $order_id)
+    {
+        $order = $this->getOrder(id: $order_id);
+
+        if ($order->getStatus() != "completed") {
+            throw new InvalidOrderStatusException($order->getStatus(), 'completed');
+        }
+
+        if (!$order->getCurrentTransaction()->isCaptured()) {
+            throw  new RefundFailedException('Order is not yet captured, only captured order could be refunded');
+        }
+
+        return $this->api_client->refundOrder(id: $order_id, orderData: ['order_lines' => $order->getOrderLines()->toArray()]);
+    }
+
+    /**
+     * Send POST request for API to create an order resource.
+     * If order data is invalid `InvalidOrderDataException` will be thrown
+     *
+     * @param Order $order
+     * @return Order
+     * @throws \GingerPluginSdk\Exceptions\InvalidOrderDataException
+     * @throws \GingerPluginSdk\Exceptions\APIException
+     */
+    public function sendOrder(Order $order): Order
     {
         try {
             $response = $this->api_client->createOrder($order->toArray());