+++ src/Entities/Transaction.php
@@ -11,6 +11,7 @@ use GingerPluginSdk\Interfaces\MultiFieldsEntityInterface;
 use GingerPluginSdk\Interfaces\ValidateFieldsInterface;
 use GingerPluginSdk\Properties\Currency;
 use GingerPluginSdk\Properties\Status;
+use SebastianBergmann\FileIterator\Facade;
 
 
 final class Transaction implements MultiFieldsEntityInterface
@@ -21,103 +22,34 @@ final class Transaction implements MultiFieldsEntityInterface
     private BaseField $paymentMethod;
     private MultiFieldsEntityInterface $paymentMethodDetails;
     private BaseField $id;
-    private BaseField $merchant_id;
-    private BaseField $created;
-    private BaseField $modified;
-    private BaseField $completed;
-    private BaseField $settled;
-    private BaseField $finalized;
-    private BaseField $expirationPeriod;
-    private BaseField $amount;
-    private BaseField $balance;
-    private BaseField $expiration_period;
-    private BaseField $description;
-    private BaseField $productType;
-    private BaseField $creditDebit;
-    private BaseField $paymentMethodBrand;
     private BaseField $paymentUrl;
     private BaseField $reason;
-    private BaseField $orderId;
     private BaseField $isCapturable;
-    private BaseField $channel;
-    private BaseField $projectType;
-    private BaseField $project_id;
-    private BaseField $captureMode;
     private BaseField $isFullyCaptured;
-    private BaseField $mscAmount;
-    private BaseField $mscCalculationType;
-    private BaseField $mscCurrency;
-    private BaseField $mscVatAmount;
-    private BaseField $mscVatClass;
-    private BaseField $mscVatPercentage;
-    private BaseField $payoutSystem;
-    private BaseField $transactionType;
 
     /**
      * @param string $paymentMethod
      * @param PaymentMethodDetails|null $paymentMethodDetails
      * @param string|null $id
-     * @param string|null $merchantId
-     * @param string|null $created
-     * @param string|null $modified
-     * @param string|null $completed
-     * @param string|null $settled
-     * @param string|null $finalized
-     * @param string|null $expirationPeriod
-     * @param \GingerPluginSdk\Properties\Currency|null $currency
-     * @param float|null $amount
-     * @param string|null $balance
-     * @param string|null $description
-     * @param string|null $productType
-     * @param string|null $creditDebit
-     * @param string|null $paymentMethodBrand
      * @param string|null $paymentUrl
      * @param \GingerPluginSdk\Properties\Status|null $status
      * @param string|null $reason
      * @param bool|null $isCapturable
-     * @param string|null $orderId
-     * @param string|null $channel
-     * @param string|null $projectType
      * @param \GingerPluginSdk\Collections\Events|null $events
-     * @param string|null $projectId
+     * @param bool|null $isFullyCaptured
+     * @param mixed ...$additionalProperties
      */
     public function __construct(
         string               $paymentMethod,
         PaymentMethodDetails $paymentMethodDetails = null,
         ?string              $id = null,
-        ?string              $merchantId = null,
-        ?string              $created = null,
-        ?string              $modified = null,
-        ?string              $completed = null,
-        ?string              $settled = null,
-        ?string              $finalized = null,
-        ?string              $expirationPeriod = null,
-        private ?Currency    $currency = null,
-        ?float               $amount = null,
-        ?string              $balance = null,
-        ?string              $description = null,
-        ?string              $productType = null,
-        ?string              $creditDebit = null,
-        ?string              $paymentMethodBrand = null,
         ?string              $paymentUrl = null,
         private ?Status      $status = null,
         ?string              $reason = null,
         ?bool                $isCapturable = null,
-        ?string              $orderId = null,
-        ?string              $channel = null,
-        ?string              $projectType = null,
         private ?Events      $events = null,
-        ?string              $projectId = null,
-        ?string              $captureMode = null,
-        ?bool                $isFullyCaptured = null,
-        ?int                 $mscAmount = null,
-        ?string              $mscCalculationType = null,
-        ?string              $mscCurrency = null,
-        ?int                 $mscVatAmount = null,
-        ?string              $mscVatClass = null,
-        ?int                 $mscVatPercentage = null,
-        ?string              $payoutSystem = null,
-        ?string              $transactionType = null,
+        ?bool                $isFullyCaptured = false,
+        mixed                ...$additionalProperties
 
     )
     {
@@ -144,127 +76,9 @@ final class Transaction implements MultiFieldsEntityInterface
         );
         $this->paymentMethodDetails = $paymentMethodDetails ?: new PaymentMethodDetails();
 
-        if ($payoutSystem) $this->payoutSystem = $this->createSimpleField(
-            propertyName: 'payout_system',
-            value: $payoutSystem
-        );
-
-        if ($transactionType) $this->transactionType = $this->createSimpleField(
-            propertyName: 'transaction_type',
-            value: $transactionType
-        );
-
-        if ($mscVatPercentage) $this->mscVatPercentage = $this->createSimpleField(
-            propertyName: 'msc_vat_percentage',
-            value: $mscVatPercentage
-        );
-
-        if ($mscVatClass) $this->mscVatClass = $this->createSimpleField(
-            propertyName: "msc_vat_class",
-            value: $mscVatClass
-        );
-
-        if ($mscVatAmount) $this->mscVatAmount = $this->createSimpleField(
-            propertyName: 'msc_vat_amount',
-            value: $mscVatAmount
-        );
-
-        if ($mscCurrency) $this->mscCurrency = $this->createSimpleField(
-            propertyName: 'msc_currency',
-            value: $mscCurrency
-        );
-
-        if ($mscCalculationType) $this->mscCalculationType = $this->createEnumeratedField(
-            propertyName: 'msc_calculation_type',
-            value: $mscCalculationType,
-            enum: ["blended", "plus", "external"]
-        );
-
-        if ($mscAmount) $this->mscAmount = $this->createFieldWithDiapasonOfValues(
-            propertyName: 'msc_amount',
-            value: $mscAmount,
-            min: 0
-        );
-
         if ($id) $this->id = $this->createSimpleField(
             'id',
             $id
-        );
-
-        if ($merchantId) $this->merchant_id = $this->createSimpleField(
-            'merchant_id',
-            $merchantId
-        );
-
-        if ($created) $this->created = $this->createFieldInDateTimeISO8601(
-            propertyName: 'created',
-            value: $created
-        );
-
-        if ($modified) $this->modified = $this->createFieldInDateTimeISO8601(
-            propertyName: 'modified',
-            value: $modified
-        );
-
-        if ($completed) $this->completed = $this->createFieldInDateTimeISO8601(
-            propertyName: 'completed',
-            value: $completed
-        );
-
-        if ($settled) $this->settled = $this->createFieldInDateTimeISO8601(
-            propertyName: 'settled',
-            value: $settled
-        );
-
-        if ($finalized) $this->finalized = $this->createFieldInDateTimeISO8601(
-            propertyName: 'finalized',
-            value: $finalized
-        );
-
-        if ($expirationPeriod) $this->expiration_period = $this->createFieldForTimePeriodISO8601(
-            propertyName: 'expiration_period',
-            value: $expirationPeriod
-        );
-
-        if ($amount) $this->amount = $this->createFieldWithDiapasonOfValues(
-            propertyName: 'amount',
-            value: $this->calculateValueInCents($amount),
-            min: 1,
-            max: 999999999999
-        );
-
-        if ($balance) $this->balance = $this->createEnumeratedField(
-            propertyName: 'balance',
-            value: $balance,
-            enum: [
-                "internal",
-                "external",
-                "test"
-            ]
-        );
-
-        if ($description) $this->description = $this->createSimpleField(
-            propertyName: 'description',
-            value: $description
-        );
-
-        if ($productType) $this->productType = $this->createSimpleField(
-            propertyName: 'product_type',
-            value: $productType
-        );
-
-        if ($creditDebit) $this->creditDebit = $this->createEnumeratedField(
-            propertyName: 'credit_debit',
-            value: $creditDebit,
-            enum: [
-                "credit",
-                "debit"
-            ]
-        );
-
-        if ($paymentMethod) $this->paymentMethodBrand = $this->createSimpleField(
-            propertyName: 'payment_method_brand',
-            value: $paymentMethodBrand
         );
 
         if ($paymentUrl) $this->paymentUrl = $this->createSimpleField(
@@ -284,46 +98,12 @@ final class Transaction implements MultiFieldsEntityInterface
             );
         }
 
-        if ($orderId) $this->orderId = $this->createSimpleField(
-            propertyName: 'order_id',
-            value: $orderId
-        );
-
-        if ($projectType) $this->projectType = $this->createSimpleField(
-            propertyName: 'project_type',
-            value: $projectType
-        );
-
-        if ($channel) $this->channel = $this->createEnumeratedField(
-            propertyName: 'channel',
-            value: $channel,
-            enum: [
-                "ecom",
-                "pos",
-                "system"
-            ]
-        );
-
-        if ($projectId) $this->project_id = $this->createSimpleField(
-            propertyName: 'project_id',
-            value: $projectId
-        );
-
-        if ($captureMode) $this->captureMode = $this->createEnumeratedField(
-            propertyName: 'capture_mode',
-            value: $captureMode,
-            enum: [
-                'manual',
-                'delayed'
-            ]
-        );
-
         if ($isFullyCaptured) $this->isFullyCaptured = $this->createSimpleField(
             propertyName: 'is_fully_captured',
             value: $isFullyCaptured
         );
 
-
+        if ($additionalProperties) $this->filterAdditionalProperties($additionalProperties);
     }
 
     public function getId(): string|bool
@@ -337,6 +117,15 @@ final class Transaction implements MultiFieldsEntityInterface
             return $this->isCapturable->get();
         }
         return false;
+    }
+
+    public function isCaptured(): bool
+    {
+        if (isset($this->isFullyCaptured)) {
+            return $this->isFullyCaptured->get();
+        } else {
+            return false;
+        }
     }
 
     public function getPaymentMethodDetails(): PaymentMethodDetails|MultiFieldsEntityInterface