+++ src/Exceptions/EmptyApiKeyException.php
@@ -5,4 +5,5 @@ namespace GingerPluginSdk\Exceptions;
 use ValueError;
 
 class EmptyApiKeyException extends ValueError
-{}
+{
+}