+++ src/Interfaces/ValueInCentsInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace GingerPluginSdk\Interfaces;
+
+interface ValueInCentsInterface
+{
+    function calculateValueInCents($value);
+}